+++ src/tqec/interop/collada/test_files/rotated_cnot.dae
